+++ Package.swift
@@ -20,6 +20,7 @@ let package = Package(
     dependencies: [
         .package(url: "https://github.com/ilyathewhite/FoundationEx.git", exact: "1.0.15"),
         .package(url: "https://github.com/ilyathewhite/AsyncNavigation.git", exact: "2.0.0"),
+        .package(url: "https://github.com/apple/swift-async-algorithms.git", exact: "1.1.5"),
         .package(url: "https://github.com/pointfreeco/swift-tagged.git", exact: "0.10.0"),
         .package(url: "https://github.com/ilyathewhite/GraphStorage.git", exact: "1.0.1")
     ],
@@ -29,6 +30,7 @@ let package = Package(
             dependencies: [
                 "FoundationEx",
                 "AsyncNavigation",
+                .product(name: "AsyncAlgorithms", package: "swift-async-algorithms"),
                 .product(name: "Tagged", package: "swift-tagged"),
                 "GraphStorage"
             ],